Merging changes from integration to main for 0.9.26.1 release

+++ cli/src/main/scala/filodb.cli/CliMain.scala
@@ -197,7 +197,7 @@ object CliMain extends StrictLogging {
             printf("%40s %20s %20s %15s %15s\n", "Child", "TotalTimeSeries", "ActiveTimeSeries", "Children", "Children")
             printf("%40s %20s %20s %15s %15s\n", "Name", "Count", "Count", "Count", "Quota")
             println("==============================================================================================================================")
-            crs._2.sortBy(c => if (addInactive) c.value.tsCount else c.value.activeTsCount)(Ordering.Int.reverse)
+            crs._2.sortBy(c => if (addInactive) c.value.tsCount else c.value.activeTsCount)(Ordering.Long.reverse)
               .foreach { cr =>
                 printf("%40s %20d %20d %15d %15d\n", cr.prefix, cr.value.tsCount,
                        cr.value.activeTsCount, cr.value.childrenCount, cr.value.childrenQuota)
